8371637: allocateNativeInternal sometimes return incorrectly aligned memory
Backported for JDK 25 on BSD.

While this fix is scheduled for inclusion in JDK 26 by upatream, I think
it's worth including in JDK 25 for BSD as it affects FreeBSD (and
possibly NetBSD), and since JDK 25 is a LTS release. Waiting for the
next LTS release for this fix seems a bit excessive.

Co-authored-by: Kurt Miller <kurt@openjdk.org>
Reviewed-by: mcimadamore, jvernee